backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
python-multipart==0.0.6
openpyxl==3.1.2
pandas==2.1.3
python-dotenv==1.0.0
cors==1.0.1 